Don't copy previous stats on failure if today has stats
There was a bug when copying previous day stats for failed nodes. If the copy had already been done, another entry was added. This duplicated the amoung of nodes in the network for this day.

For some reason some old hubzilla nodes are registering up to ~40 times at some point in the night and then returning an error.