Remove charts from protocol pages for now.
Reason is that protocol is changing over time
thing. We can't for example count all mastodon servers towards activitypub since the
support for that is recent and even not landing on all servers at the same time. Some
platforms allow turning protocols on and off. If we want to show history charts, we
must store protocol history per node first.